be more explicit about types

+++ contracts/nzcp.sol
@@ -23,14 +23,14 @@ contract NZCP is EllipticCurve {
     // "key-1" public key published here:
     // https://nzcp.covid19.health.nz/.well-known/did.json
     // Doesn't suppose to change unless MoH leaks their private key
-    uint public constant EXAMPLE_X = 0xCD147E5C6B02A75D95BDB82E8B80C3E8EE9CAA685F3EE5CC862D4EC4F97CEFAD;
-    uint public constant EXAMPLE_Y = 0x22FE5253A16E5BE4D1621E7F18EAC995C57F82917F1A9150842383F0B4A4DD3D;
+    uint256 public constant EXAMPLE_X = 0xCD147E5C6B02A75D95BDB82E8B80C3E8EE9CAA685F3EE5CC862D4EC4F97CEFAD;
+    uint256 public constant EXAMPLE_Y = 0x22FE5253A16E5BE4D1621E7F18EAC995C57F82917F1A9150842383F0B4A4DD3D;
 
     // "z12Kf7UQ" public key published here:
     // https://nzcp.identity.health.nz/.well-known/did.json
     // Doesn't suppose to change unless MoH leaks their private key
-    uint public constant LIVE_X = 0x0D008A26EB2A32C4F4BBB0A3A66863546907967DC0DDF4BE6B2787E0DBB9DAD7;
-    uint public constant LIVE_Y = 0x971816CEC2ED548F1FA999933CFA3D9D9FA4CC6B3BC3B5CEF3EAD453AF0EC662;
+    uint256 public constant LIVE_X = 0x0D008A26EB2A32C4F4BBB0A3A66863546907967DC0DDF4BE6B2787E0DBB9DAD7;
+    uint256 public constant LIVE_Y = 0x971816CEC2ED548F1FA999933CFA3D9D9FA4CC6B3BC3B5CEF3EAD453AF0EC662;
 
     // 27 bytes to skip the ["Signature1", headers, buffer0] start of ToBeSignedBuffer
     // And get to the CWT claims straight away
@@ -203,7 +203,7 @@ contract NZCP is EllipticCurve {
                     (pos, cbor_type2, v2) = readType(buffer, pos);
                     require(cbor_type2 == MAJOR_TYPE_INT, "cbortype expected to be integer");
 
-                    uint256 exp;
+                    uint exp;
                     (pos, exp) = decodeCBORUint(buffer, pos, v2);
                     require(block.timestamp < exp, "Pass expired"); // check if pass expired
                 }
@@ -279,7 +279,7 @@ contract NZCP is EllipticCurve {
 
     // Verifies NZCP message hash signature
     // Returns true if signature is valid, reverts transaction otherwise
-    function verifySignature(bytes32 messageHash, uint[2] memory rs, bool is_example) public pure returns (bool) {
+    function verifySignature(bytes32 messageHash, uint256[2] memory rs, bool is_example) public pure returns (bool) {
         if (is_example) {
             require(validateSignature(messageHash, rs, [EXAMPLE_X, EXAMPLE_Y]), "Invalid signature");
             return true;
@@ -292,13 +292,13 @@ contract NZCP is EllipticCurve {
 
     // Verifies NZCP ToBeSignedBuffer
     // Returns true if signature is valid, reverts transaction otherwise
-    function verifyToBeSignedBuffer(bytes memory buffer, uint[2] memory rs, bool is_example) public pure returns (bool) {
+    function verifyToBeSignedBuffer(bytes memory buffer, uint256[2] memory rs, bool is_example) public pure returns (bool) {
         return verifySignature(sha256(buffer), rs, is_example);
     }
 
     // Parses ToBeSignedBuffer and returns the credential subject
     // Returns credential subject if pass is valid, reverts transaction otherwise
-    function parseAndVerifyToBeSignedBuffer(bytes memory buffer, uint[2] memory rs, bool is_example) public view 
+    function parseAndVerifyToBeSignedBuffer(bytes memory buffer, uint256[2] memory rs, bool is_example) public view 
         returns (string memory, string memory, string memory) {
 
         uint credentialSubjectPos = findCredentialSubject(buffer, claims_skip, 0);
